+++ packages/sdk/src/components/provider.tsx
@@ -1,6 +1,7 @@
 import { createContext, useContext, useMemo } from "react";
 import { createTrackHelpers } from "../api/track-helpers";
-import { type AnalyticsOptions, type AnalyticsProviderProps, type TrackHelpers } from "../types";
+import { type AnalyticsOptions, type TrackHelpers } from "../types";
+import { type AnalyticsProviderProps } from "../types/react";
 
 const emptyOptions: AnalyticsOptions = {};
 
@@ -10,13 +11,15 @@ export function AnalyticsProvider({
 	projectId,
 	ingestUrl,
 	debug,
+	path,
+	referrer,
 	children,
 }: AnalyticsProviderProps) {
 	const value = useMemo(
 		function () {
-			return { projectId, ingestUrl, debug };
+			return { projectId, ingestUrl, debug, path, referrer };
 		},
-		[projectId, ingestUrl, debug],
+		[projectId, ingestUrl, debug, path, referrer],
 	);
 
 	return <AnalyticsContext.Provider value={value}>{children}</AnalyticsContext.Provider>;
@@ -32,7 +35,7 @@ export function useTrack(): TrackHelpers {
 		function () {
 			return createTrackHelpers(options);
 		},
-		[options.projectId, options.ingestUrl, options.debug],
+		[options.projectId, options.ingestUrl, options.debug, options.path, options.referrer],
 	);
 }